+++ src/angular/core/overlay/index.ts
@@ -1,3 +1,4 @@
+export * from './focus-initial';
 export * from './overlay-base';
 export * from './overlay-config';
 export * from './overlay-container-base';